+++ www/OSSocialLogins.js
@@ -12,6 +12,10 @@ exports.loginGoogle = function (success, error, clientId, redirectUrl) {
     exec(success, error, 'OSSocialLogins', 'loginGoogle', [clientId, redirectUrl]);
 };
 
+exports.loginFacebook = function (success, error) {
+    exec(success, error, 'OSSocialLogins', 'loginFacebook');
+}
+
 exports.logout = function (arg0, success, error) {
     exec(success, error, 'OSSocialLogins', 'logout', [arg0]);
 };